apps/api/pyproject.toml
[project]
name = "oiagent-api"
version = "0.1.0"
description = "OIAgent FastAPI backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "asyncpg>=0.29",
    "supabase>=2.4",
    "langchain>=0.2",
    "langchain-google-genai>=1.0",
    "langchain-anthropic>=0.1",
    "langgraph>=0.1",
    "python-multipart>=0.0.9",
    "pypdf>=4.2",
    "openpyxl>=3.1",
    "python-docx>=1.1",
    "pyjwt>=2.8",
    "cryptography>=43.0",
]

[dependency-groups]
dev = [
    "ruff>=0.4",
    "mypy>=1.10",
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = [
    "B008",  # FastAPI Depends() used as default argument — intentional pattern
]

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
